add lock file to Bystro annotator, to ensure multiple instances of Bystro do not run and produce outputs in the same directory

+++ python/python/bystro/beanstalkd/worker.py
@@ -1,6 +1,7 @@
 """TODO: Add description here"""
 
-from concurrent.futures import ThreadPoolExecutor
+from concurrent import futures
+from concurrent.futures import ProcessPoolExecutor
 import logging
 import os
 import signal
@@ -10,7 +11,8 @@ import traceback
 from collections.abc import Callable
 from textwrap import dedent
 from typing import Any, TypeVar
-
+import psutil
+import cloudpickle
 
 from msgspec import DecodeError, ValidationError, json
 from pystalk import BeanstalkClient, BeanstalkError  # type: ignore
@@ -22,16 +24,14 @@ from bystro.beanstalkd.messages import (
     InvalidJobMessage,
     ProgressPublisher,
     QueueConf,
-    ProgressMessage
+    ProgressMessage,
 )
 
-executor = ThreadPoolExecutor(max_workers=1)
-
 BEANSTALK_ERR_TIMEOUT = "TIMED_OUT"
-QUEUE_RESERVE_JOB_TIMEOUT_TIME = int(os.getenv("BYSTRO_BEANSTALKD_RESERVE_JOB_TIMEOUT", "20"))
-QUEUE_JOB_HEARTBEAT_INTERVAL = int(os.getenv("BYSTRO_BEANSTALKD_JOB_HEARTBEAT_INTERVAL", "20"))
+QUEUE_RESERVE_JOB_TIMEOUT_TIME = int(os.getenv("BYSTRO_BEANSTALKD_RESERVE_JOB_TIMEOUT", "5"))
+QUEUE_JOB_HEARTBEAT_INTERVAL = int(os.getenv("BYSTRO_BEANSTALKD_JOB_HEARTBEAT_INTERVAL", "5"))
 # Must be larger than QUEUE_RESERVE_JOB_TIMEOUT_TIME and QUEUE_JOB_HEARTBEAT_INTERVAL
-QUEUE_CONSUMER_SOCKET_TIMEOUT = int(os.getenv("BYSTRO_BEANSTALKD_CONSUMER_SOCKET_TIMEOUT", "30"))
+QUEUE_CONSUMER_SOCKET_TIMEOUT = int(os.getenv("BYSTRO_BEANSTALKD_CONSUMER_SOCKET_TIMEOUT", "10"))
 
 T = TypeVar("T", bound=BaseMessage)
 T2 = TypeVar("T2", bound=BaseMessage)
@@ -39,16 +39,45 @@ T3 = TypeVar("T3", bound=BaseMessage)
 
 logger = logging.getLogger(__name__)
 
+class FunctionWrapper:
+    def __init__(self, fn):
+        self.fn_ser = cloudpickle.dumps(fn)
+
+    def __call__(self, *args, **kwargs):
+        fn = cloudpickle.loads(self.fn_ser)
+        return fn(*args, **kwargs)
+
+class NotFoundError(Exception):
+    """
+    A NOT_FOUND Beanstalkd error arises
+    when a job is no longer bound to the worker
+
+    In these cases, there is no safe way to continue processing the job
+    """
+
+    def __init__(self, job_id: BeanstalkJobID):
+        super().__init__(f"Job {job_id} is no longer bound to this worker")
+
 
 # Signal handler function
 def sigterm_handler(_signum, _frame):
-    print("SIGTERM received. Cleaning up...", file=sys.stderr)
-    executor.shutdown(wait=False)
-    exit(0)
+    print("SIGINT received. Cleaning up...")
+    # executor.shutdown(wait=False)
+
+    parent = psutil.Process(os.getpid())
+    print("parent", parent)
+    print("children", parent.children(recursive=True))
+    for child in parent.children(recursive=True):  # or parent.children() for recursive=False
+        print("child", child)
+        child.kill()
+    parent.kill()
+    sys.exit(0)
 
 
 # Set up the signal handler in the main thread
-signal.signal(signal.SIGTERM, sigterm_handler)
+# signal.signal(signal.SIGTERM, sigterm_handler)
+signal.signal(signal.SIGINT, sigterm_handler)
+# signal.signal(signal.SIGQUIT, sigterm_handler)
 
 
 def handle_job(handler_fn, publisher, job_data):
@@ -57,7 +86,9 @@ def handle_job(handler_fn, publisher, job_data):
     except Exception as e:
         return e
 
-
+def handler_fnBaz(publisher, job_data):
+    print("GOT JOB DATA", job_data)
+    return job_data
 def default_failed_msg_fn(
     job_data: T | None, job_id: BeanstalkJobID, err: Exception
 ) -> FailedJobMessage | InvalidJobMessage:  # noqa: E501
@@ -67,17 +98,41 @@ def default_failed_msg_fn(
     return FailedJobMessage(submission_id=job_data.submission_id, reason=str(err))
 
 
-def _touch(client: BeanstalkClient, job_id: str | int):
-    # Ping the server periodically while waiting for the handler to finish
+def _shutdown():
+    executor.shutdown(wait=False)
+
+    logger.error("Shutdown called")
+    parent = psutil.Process(os.getpid())
+    for child in parent.children(recursive=True):  # or parent.children() for recursive=False
+        logger.error(f"Killing child PID: {child}")
+        try:
+            child.kill()
+        except Exception:
+            logger.error(f"Failed to kill child PID: {child}")
     try:
-        with client._sock_ctx() as socket:  # noqa: SLF001
-            client._send_message(f"touch {job_id}", socket)  # noqa: SLF001
-            body = client._receive_word(socket, b"TOUCHED", b"NOT_FOUND")  # noqa: SLF001
-            if body == b"NOT_FOUND":
-                logger.warning("Job %s not found", job_id)
-            logger.debug("Touched job %s", job_id)
-    except Exception as e:
-        logger.exception("Ping error while waiting for handler: %s", e)
+        parent.kill()
+    except Exception:
+        logger.error(f"Failed to kill parent PID: {parent}")
+    sys.exit(0)
+
+
+def _touch(client: BeanstalkClient, job_id: BeanstalkJobID, future: Any):
+    # Ping the server periodically while waiting for the handler to finish
+    with client._sock_ctx() as socket:  # noqa: SLF001
+        client._send_message(f"touch {job_id}", socket)  # noqa: SLF001
+        return client._receive_word(socket, b"TOUCHED")  # noqa: SLF001
+
+
+print("os.getpid() in worker global", os.getpid())
+
+class WrapHandlerFn:
+    def __init__(self, handler_fn: Callable[[ProgressPublisher, T], Any], publisher: ProgressPublisher):
+        self.handler_fn = cloudpickle.dumps(handler_fn)
+        self.publisher = publisher
+
+    def __call__(self, job_data: T) -> Any:
+        handler_fn = cloudpickle.loads(self.handler_fn)
+        return handler_fn(self.publisher, job_data)
 
 
 def listen(
@@ -99,6 +154,7 @@ def listen(
     )
 
     i = 0
+    future: futures.Future | None = None
     while True:
         i += 1
         job = None
@@ -118,40 +174,30 @@ def listen(
 
             try:
                 job_data = json.decode(job.job_data, type=job_data_type)
-            except ValidationError as err:
-                msg = dedent(
-                    f"""
+            except Exception as err:
+                msg = Exception("Unknown error, check admin logs")
+    
+                if isinstance(err, DecodeError):
+                    msg = ValueError(dedent(
+                        f"""
+                            Job {job_id} JSON is invalid.
+                            Decoding `{str(job.job_data)}`, failed with: `{err}`"""
+                    ))
+                if isinstance(err, ValidationError):
+                    msg = ValueError(dedent(
+                        f"""
                             Job {job_id} JSON does not have the data expected.
                             Expected {job_data_type.keys_with_types()}.
                             Decoding failed with: `{err}`"""
-                )
-                traceback.print_exc()
-                client.put_job(json.encode(failed_msg_fn(job_data, job_id, ValueError(msg))))
-                client.delete_job(job.job_id)
-                continue
-            except DecodeError as err:
-                msg = dedent(
-                    f"""
-                            Job {job.job_id} JSON is invalid.
-                            Decoding `{str(job.job_data)}`, failed with: `{err}`"""
-                )
-                traceback.print_exc()
-                client.put_job(json.encode(failed_msg_fn(job_data, job_id, ValueError(msg))))
-                client.delete_job(job_id)
-                continue
-            except Exception:
-                traceback.print_exc()
+                    ))
 
-                client.put_job(
-                    json.encode(
-                        failed_msg_fn(job_data, job_id, Exception("Unknown error, check admin logs"))
-                    )
-                )
+                client.put_job(json.encode(failed_msg_fn(job_data, job_id, ValueError(msg))))
                 client.delete_job(job_id)
+
+                traceback.print_exc()
+                continue
 
             try:
-                assert job_data is not None
-
                 publisher = ProgressPublisher(
                     host=client.host,
                     port=client.port,
@@ -160,33 +206,62 @@ def listen(
                 )
 
                 client.put_job(json.encode(submit_msg_fn(job_data)))
+                with ProcessPoolExecutor(max_workers=1) as executor:
+                    future = executor.submit(FunctionWrapper(handler_fn), publisher, job_data)
 
-                # Submit the job to the ThreadPoolExecutor
-                future = executor.submit(handle_job, handler_fn, publisher, job_data)
+                    # Ensure job is kept alive indefinitely, until completion
+                    # Some jobs are potentially weeks long
+                    last_touch_time = time.time()
 
-                # Ensure job is kept alive indefinitely, until completion
-                # Some jobs are potentially weeks long
-                last_touch_time = time.time()
-                while True:
-                    # Check if the handle_job task is complete
-                    if future.done():
-                        _touch(client, job_id)
-                        last_touch_time = time.time()
-                        break
+                    failed_touch = False
+                    while True:
+                        failed_touch = False
+                        # Check if the handle_job task is complete
+                        try:
+                            if future.done():
+                                _touch(client, job_id, future)
+                                last_touch_time = time.time()
+                                break
 
-                    if time.time() - last_touch_time >= QUEUE_JOB_HEARTBEAT_INTERVAL:
-                        _touch(client, job_id)
-                        last_touch_time = time.time()
+                            if time.time() - last_touch_time >= QUEUE_JOB_HEARTBEAT_INTERVAL:
+                                _touch(client, job_id, future)
+                                last_touch_time = time.time()
+                        except TimeoutError:
+                            logger.warning("Job %s  _touch timed out", job_id)
+                        except Exception as err:
+                            # The only expected error is NOT_FOUND which means the job is no longer bound
+                            # to the worker.
+                            # This means we must terminate processing of this job, since other
+                            # workers may to pick it up for processing
+                            # Any other exception is even more odd, and suggest a fatal error
+                            # related to beanstalkd communication, which again could potentially
+                            # lead to duplicate job processing
+                            logger.error(
+                                "Job %s _touch failed with %s, shutting down job processing", job_id, err
+                            )
+                            failed_touch = True
+                            break
 
-                    time.sleep(1)
+                    if failed_touch:
+                        future.cancel()
+                        executor.shutdown(wait=True, cancel_futures=True)
+                        failed_touch = True
 
-                res = future.result()
+                        try:
+                            client.release_job(job.job_id)
+                        except Exception as err:
+                            logger.error("Failed to release job with id %s due to: %s", job.job_id, err)
 
-                if isinstance(res, Exception):
-                    raise res
+                        continue
 
-                client.put_job(json.encode(completed_msg_fn(job_data, res)))
-                client.delete_job(job.job_id)
+                    res = future.result()
+                    print("res", res)
+
+                    if isinstance(res, Exception):
+                        raise res
+
+                    client.put_job(json.encode(completed_msg_fn(job_data, res)))
+                    client.delete_job(job.job_id)
             except Exception as err:
                 traceback.print_exc()
 
@@ -201,6 +276,8 @@ def listen(
             if err.message == BEANSTALK_ERR_TIMEOUT:
                 continue
 
+            logger.warning("Beanstalkd error: %s")
+
             traceback.print_exc()
 
             if client is None:
@@ -213,6 +290,22 @@ def listen(
                 time.sleep(1)
                 continue
 
+            logger.warning("Releasing job with id %s", job.job_id)
+
             client.release_job(job.job_id)
             time.sleep(1)
             continue
+        # Other exceptions are unexpected and should result in a restart to ensure all state is cleared
+        # Restart is managed by the container orchestrator or process manager
+        except Exception as err:
+            logger.exception("Unknown error: %s", err)
+
+            if client is not None and job is not None:
+                logger.error("Attempting to release job with id %s due to: %s", err. job.job_id)
+
+                try:
+                    client.release_job(job.job_id)
+                except Exception as err:
+                    logger.error("Failed to release job with id %s due to: %s", job.job_id, err)
+
+            time.sleep(1)
